Changes relevant to the last pull request
