+++ src/Components/Actions/ActionsTab.css
@@ -2,7 +2,7 @@
   background-color: sandybrown;
 }
 
-#moreActionsBtn{
+#actionsWindowBtn{
   background-color: #ffa500; 
   padding: 0.5rem; 
   width: 50%;
