+++ src/Components/Actions/ActionsWindow.js
@@ -2,17 +2,17 @@ import React from 'react'
 import GeneralWindow from 'Components/GeneralComponents/GeneralWindow';
 
 export default function ActionsWindow({ unmountActionsWindow }) {
-    return (
-        <GeneralWindow unmountGeneralWindow={unmountActionsWindow}>
-        <br />
-        <br />
-        <br />
+  return (
+    <GeneralWindow unmountGeneralWindow={unmountActionsWindow}>
+      <br />
+      <br />
+      <br />
         Actions Window
-        <br />
-        <br />
-        <br />
-        <br />
+      <br />
+      <br />
+      <br />
+      <br />
         Actions....
-      </GeneralWindow>
-    )
+    </GeneralWindow>
+  )
 }
