--- src/Utils/GeneralComponents/useCountdown.js
@@ -1,30 +1,0 @@
-import { useState, useEffect } from "react";
-import moment from "moment";
-import "moment-duration-format";
-export default function useCountdown(startTime, endTime) {
-  const [duration, setDuration] = useState(
-    moment.duration(endTime - startTime)
-  );
-  const [hasReachedZero , setHasReachedZero ] = useState(false);
-  useEffect(() => {
-    const interval = 1000;
-
-    let intervalDuration;
-    if (!hasReachedZero) {
-      intervalDuration = setInterval(() => {
-        setDuration(moment.duration(duration.asMilliseconds() - interval));
-      }, interval);
-    }
-
-    if (!hasReachedZero && duration.asMilliseconds() <= 0) {
-      setDuration(moment.duration(0));
-      setHasReachedZero(true);
-      clearInterval(intervalDuration);
-    }
-    return () => {
-      clearInterval(intervalDuration);
-    };
-  }, [duration, hasReachedZero]);
-
-  return duration.asMilliseconds();
-}
